+++ packages/dev-bot/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.13.3"></a>
+## [10.13.3](https://github.com/botpress/botpress/compare/v10.13.2...v10.13.3) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package botpress-dev-bot
+
+<a name="10.13.2"></a>
+## [10.13.2](https://github.com/botpress/botpress/compare/v10.13.1...v10.13.2) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package botpress-dev-bot
+
+<a name="10.13.1"></a>
+## [10.13.1](https://github.com/botpress/botpress/compare/v10.13.0...v10.13.1) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package botpress-dev-bot
+
+<a name="10.13.0"></a>
+# [10.13.0](https://github.com/botpress/botpress/compare/v10.11.3...v10.13.0) (2018-06-01)
+
+
+### Bug Fixes
+
+* **qna:** properly process the redirect node ([d513f24](https://github.com/botpress/botpress/commit/d513f24))
+
+
+### Features
+
+* **qna:** Q&A module ([f79e2e8](https://github.com/botpress/botpress/commit/f79e2e8))
+* **qna:** redirect to node, middleware not working yet ([d93949b](https://github.com/botpress/botpress/commit/d93949b))
+
+
+
+
+<a name="10.12.0"></a>
+# [10.12.0](https://github.com/botpress/botpress/compare/v10.11.3...v10.12.0) (2018-06-01)
+
+
+### Bug Fixes
+
+* **qna:** properly process the redirect node ([d513f24](https://github.com/botpress/botpress/commit/d513f24))
+
+
+### Features
+
+* **qna:** Q&A module ([f79e2e8](https://github.com/botpress/botpress/commit/f79e2e8))
+* **qna:** redirect to node, middleware not working yet ([d93949b](https://github.com/botpress/botpress/commit/d93949b))
+
+
+
+
 <a name="10.11.1"></a>
 ## [10.11.1](https://github.com/botpress/botpress/compare/v10.11.0...v10.11.1) (2018-05-29)