+++ packages/functionals/botpress-nlu/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.13.3"></a>
+## [10.13.3](https://github.com/botpress/modules/compare/v10.13.2...v10.13.3) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @botpress/nlu
+
+<a name="10.13.2"></a>
+## [10.13.2](https://github.com/botpress/modules/compare/v10.13.1...v10.13.2) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @botpress/nlu
+
+<a name="10.13.1"></a>
+## [10.13.1](https://github.com/botpress/modules/compare/v10.13.0...v10.13.1) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @botpress/nlu
+
+<a name="10.13.0"></a>
+# [10.13.0](https://github.com/botpress/modules/compare/v10.11.3...v10.13.0) (2018-06-01)
+
+
+### Features
+
+* **nlu:** hide hidden intents by default ([077da1f](https://github.com/botpress/modules/commit/077da1f))
+* **qna:** Q&A module ([f79e2e8](https://github.com/botpress/modules/commit/f79e2e8))
+
+
+
+
+<a name="10.12.0"></a>
+# [10.12.0](https://github.com/botpress/modules/compare/v10.11.3...v10.12.0) (2018-06-01)
+
+
+### Features
+
+* **nlu:** hide hidden intents by default ([077da1f](https://github.com/botpress/modules/commit/077da1f))
+* **qna:** Q&A module ([f79e2e8](https://github.com/botpress/modules/commit/f79e2e8))
+
+
+
+
 <a name="10.11.1"></a>
 ## [10.11.1](https://github.com/botpress/modules/compare/v10.11.0...v10.11.1) (2018-05-29)